refactor: reorganize interface files
